refactor(submission): combine submission example
